Develop

+++ src/main/java/com/gym/dao/FacturaDAO.java
@@ -17,54 +17,56 @@ public class FacturaDAO {
 	public FacturaDAO(Connection con) {
 		this.con = con;
 	}
-	public List<Factura> consultarFactura(){
-		List<Factura> resultado = new ArrayList<>();
-//		int elementosPorPagina = 10; 
-//		int paginaActual = 1;
+	
+
+	public Factura consultarFactura(int id){
+		Factura factura = null;
 		
 		try {
 			
-			String sentencia = "select * from factura f, usuario u where f.administrador_id = ? and f.usuario_id = u.id";
+			String sentencia = "select * from factura f, usuario u where f.usuario_id = u.id and f.id = ?";
 			
 			
 			final PreparedStatement statement = con.prepareStatement(sentencia);
 			
 			try(statement) {
+				statement.setInt(1, id);
 				
 				final ResultSet resultSet = statement.executeQuery();
 				
 				try(resultSet) {
 					
-					while(resultSet.next()) {
-						resultado.add(new Factura(
-			                    resultSet.getInt("f.id"), 
-			                    resultSet.getString("f.numero_factura"),
-			                    resultSet.getDouble("f.descuento_porcentaje"),
-			                    resultSet.getDouble("f.descuento"),
-			                    resultSet.getDouble("f.subtotal"),
-			                    resultSet.getDouble("f.iva"),
-			                    resultSet.getDouble("f.total"),
-			                    resultSet.getString("f.forma_pago"),
-			                    resultSet.getString("f.fecha"),
-			                    resultSet.getString("f.establecimiento"),
-			                    resultSet.getString("f.punto_emision"),
-			                    resultSet.getInt("f.usuario_id"),
-			                    resultSet.getInt("f.administrador_id"),
-			                    resultSet.getString("u.nombre")
-								));
-					}
+					resultSet.next();
 					
-					return resultado;
+					factura = new Factura(
+	                    resultSet.getInt("f.id"), 
+	                    resultSet.getString("f.numero_factura"),
+	                    resultSet.getDouble("f.descuento_porcentaje"),
+	                    resultSet.getDouble("f.descuento"),
+	                    resultSet.getDouble("f.subtotal"),
+	                    resultSet.getDouble("f.iva"),
+	                    resultSet.getDouble("f.total"),
+	                    resultSet.getString("f.forma_pago"),
+	                    resultSet.getDate("f.fecha"),
+	                    resultSet.getString("f.establecimiento"),
+	                    resultSet.getString("f.punto_emision"),
+	                    resultSet.getInt("f.usuario_id"),
+	                    resultSet.getInt("f.administrador_id"),
+	                    resultSet.getString("u.nombre"),
+	                    resultSet.getString("u.apellido"),
+	                    resultSet.getString("u.cedula")
+							);
 					
+					return factura;
 				}
-				
 			}
 			
 		} catch(SQLException e) {
 			e.printStackTrace();
-			return resultado;
+			return factura;
 		}
 	}
+
 	
 	public List<Factura> listarFactura(int administrador_id, String nombre) {
 		
@@ -74,10 +76,10 @@ public class FacturaDAO {
 		
 		try {
 			
-			String sentencia = "select * from factura f, usuario u where f.administrador_id = ? and f.usuario_id = u.id";
+			String sentencia = "select * from factura f, usuario u where f.administrador_id = ? and f.usuario_id = u.id and total <> 0.0";
 			
 			if(!nombre.equals("")) {
-				sentencia = "select * from factura f, usuario u where f.administrador_id = ? and f.usuario_id = u.id and nombre like ?";
+				sentencia = "select * from factura f, usuario u where f.administrador_id = ? and f.usuario_id = u.id and nombre like ? and total <> 0.0";
 			}
 			
 			final PreparedStatement statement = con.prepareStatement(sentencia);
@@ -103,7 +105,7 @@ public class FacturaDAO {
 			                    resultSet.getDouble("f.iva"),
 			                    resultSet.getDouble("f.total"),
 			                    resultSet.getString("f.forma_pago"),
-			                    resultSet.getString("f.fecha"),
+			                    resultSet.getDate("f.fecha"),
 			                    resultSet.getString("f.establecimiento"),
 			                    resultSet.getString("f.punto_emision"),
 			                    resultSet.getInt("f.usuario_id"),
@@ -150,7 +152,9 @@ public class FacturaDAO {
 		Factura factura = null;
 		
 		try {
-			String sentencia = "select * from factura where id = last_insert_id() and administrador_id = ?";
+			String sentencia = "select id, numero_factura, descuento_porcentaje, descuento, subtotal,"
+					+ " iva, total, forma_pago, date(fecha) as fecha, establecimiento, punto_emision, usuario_id, administrador_id"
+					+ " from factura where id = maxIdFactura() and administrador_id = ?";
 			
 			final PreparedStatement statement = con.prepareStatement(sentencia);
 			
@@ -160,11 +164,14 @@ public class FacturaDAO {
 				final ResultSet resultSet = statement.executeQuery();
 				
 				try(resultSet) {
+					
+					resultSet.next();
+					
 					return factura = new Factura(
 							resultSet.getInt("id"),
 							resultSet.getString("numero_factura"), 
 							resultSet.getString("forma_pago"), 
-							resultSet.getString("fecha"), 
+							resultSet.getDate("fecha"), 
 							resultSet.getString("establecimiento"), 
 							resultSet.getString("punto_emision") 
 							);
@@ -178,26 +185,85 @@ public class FacturaDAO {
 		}
 	}
 	
+	public int consultarUsuarioUltimaFactura(int administrador_id) {
+		
+		try {
+			String sentencia = "select usuario_id from factura where id = maxIdFactura() and administrador_id = ?";
+			
+			final PreparedStatement statement = con.prepareStatement(sentencia);
+			
+			try(statement){
+				statement.setInt(1, administrador_id);
+				
+				final ResultSet resultSet = statement.executeQuery();
+				
+				try(resultSet) {
+					
+					resultSet.next();
+					
+					if(Utilidades.isNumber(resultSet.getInt("usuario_id") + "")) {
+						return resultSet.getInt("usuario_id");
+					} else {
+						return 0;
+					}
+				}
+	
+			}
+		}catch(SQLException e) {
+			e.printStackTrace();
+			return 0;
+		}
+	}
+	
+	// Verificamos si la ultima factura creada tiene productos para saber si tenemos que crear otra o podemos usar la misma
+	public boolean ultimaFacturaIncompleta(int administrador_id) {
+		
+		try {
+			String sentencia = "select * from factura where id = maxIdFactura() and administrador_id = ? and total = 0.00";
+			
+			final PreparedStatement statement = con.prepareStatement(sentencia);
+			
+			try(statement){
+				statement.setInt(1, administrador_id);
+				
+				final ResultSet resultSet = statement.executeQuery();
+				
+				try(resultSet) {
+					
+					if(resultSet.next()) {
+						return true;
+					} else {
+						return false;
+					}
+					
+				}
+	
+			}
+		}catch(SQLException e) {
+			e.printStackTrace();
+			return false;
+		}
+	}
+	
 	public boolean actualizarFactura(Factura factura) {
 		int item = 0;
 		
 		try {
-			String sentencia = "update factura set descuento_porcentaje = ?, descuento = ?, subtotal = ?, iva = ?, total = ?,"
-					+ " forma_pago = ?, fecha = ?, usuario_id = ? where id = ?";
+			String sentencia = "call actualizarFactura(?,?,?,?,?,?,?,?,?)";
 			
 			final PreparedStatement statement = con.prepareStatement(sentencia);
 			
 			try(statement){
 				
-				statement.setDouble(1, factura.getDescuento_porcentaje());
-				statement.setDouble(1, factura.getDescuento());
-				statement.setDouble(1, factura.getSubtotal());
-				statement.setDouble(1, factura.getIva());
-				statement.setDouble(1, factura.getTotal());
-				statement.setString(1, factura.getForma_pago());
-				statement.setString(1, factura.getFecha());
-				statement.setInt(1, factura.getUsuario_id());
 				statement.setInt(1, factura.getId());
+				statement.setDouble(2, factura.getDescuento_porcentaje());
+				statement.setDouble(3, factura.getDescuento());
+				statement.setDouble(4, factura.getSubtotal());
+				statement.setDouble(5, factura.getIva());
+				statement.setDouble(6, factura.getTotal());
+				statement.setString(7, factura.getForma_pago());
+				statement.setDate(8, factura.getFecha());
+				statement.setInt(9, factura.getUsuario_id());
 				
 				item = statement.executeUpdate();
 	
@@ -210,6 +276,10 @@ public class FacturaDAO {
 	}
 	
 	public int eliminarFactura(int id) {
+		
+		if(!eliminarMembresiasFactura(id)) {
+			return 0;
+		}
 		
 		try {
 			String sentencia = "delete from factura where id = ?";
@@ -229,4 +299,26 @@ public class FacturaDAO {
 		}
 		
 	}
+	
+	public boolean eliminarMembresiasFactura(int factura_id) {
+		
+		try {
+			String sentencia = "delete from membresia where factura_id = ?";
+			
+			final PreparedStatement statement = con.prepareStatement(sentencia);
+			
+			try(statement){
+				
+				statement.setInt(1, factura_id);
+				
+				return new Utilidades().toBoolean(statement.executeUpdate());
+	
+			}
+		}catch(SQLException e) {
+			e.printStackTrace();
+			return false;
+		}
+		
+	}
+	
 }
